Project File/requirements.txt
Flask==2.3.3
spacy==3.5.0
pdfplumber==0.10.3
python-docx==1.1.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
transformers==4.30.0
torch==2.0.1
werkzeug==2.3.7
pydantic==1.10.12  # Fixed version to avoid the REGEX error